Only try to delete zip if it exists, fix ambiguous use of "replace"

+++ bin/index.js
@@ -93,10 +93,10 @@ function combineSizes(sizes, sizeObj) {
 	return sizes;
 }
 
-function replaceVals(search, replace, paths) {
+function replaceVals(search, changeTo, paths) {
 	replace({
 		regex: search,
-		replacement: replace,
+		replacement: changeTo,
 		paths: paths
 	});
 }
@@ -117,6 +117,7 @@ function performReplacements() {
 }
 
 function replaceTestJS() {
+	replaceVals(oldSizes.width + 'x' + oldSizes.height, newSizes.width + 'x' + newSizes.height, testPath);
 	replaceVals(/\.width\)\.not\.toBeGreaterThan\(\d+\)/, '.width).not.toBeGreaterThan(' + newSizes.width + ')', testPath);
 	replaceVals(/\.height\)\.not\.toBeGreaterThan\(\d+\)/, '.height).not.toBeGreaterThan(' + newSizes.height + ')', testPath);
 }
@@ -170,6 +171,6 @@ function deleteZip() {
 			return /\.zip$/.test(file);
 		})[0];
 
-	fs.unlinkSync('./' + zipPath);
+	if (zipPath) fs.unlinkSync('./' + zipPath);
 }
